% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_overlap_to_edges}
\alias{add_overlap_to_edges}
\title{Add overlap information to edges}
\usage{
add_overlap_to_edges(g, gene_sets)
}
\arguments{
\item{g}{The pathway graph (igraph object).}

\item{gene_sets}{A named list of gene sets.}
}
\value{
The graph with overlap information added to edges.
}
\description{
Add overlap information to edges
}
