% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph_functions.R
\name{create_glyph_on_the_fly}
\alias{create_glyph_on_the_fly}
\title{Create a glyph for a pathway on-the-fly}
\usage{
create_glyph_on_the_fly(
  pathway,
  conditions,
  enrichment_scores,
  enrichment_limits,
  glyph_size = c(80, 60),
  res = 96,
  lollipop_plot = T,
  lollipop_colors = NULL
)
}
\arguments{
\item{pathway}{Character string representing the pathway name.}

\item{conditions}{Character vector of conditions to include in the glyph.}

\item{enrichment_scores}{Data frame of enrichment scores (rows are pathways, columns are conditions).}

\item{enrichment_limits}{Numeric vector of length 2 specifying the enrichment limits.}

\item{glyph_size}{Numeric vector specifying the width and height of the glyph in pixels. Default is \code{c(80, 60)}.}

\item{res}{Numeric value specifying the resolution of the image in dpi. Default is 96.}

\item{lollipop_plot}{Bool specifying whether we should create lollipop plots (\code{TRUE}) or not (\code{FALSE}). Lollipop plots make large numbers of comparisons easier to compare than barplots, especially when combined with lollipop_colors, with a slight loss in enrichment strength fidelity.}

\item{lollipop_colors}{Optional named list, specifies the colors of conditions in the lollipop plot. Names must match condititions exactly}
}
\value{
A magick image representing the glyph for the pathway as well as the local path to the image.
}
\description{
Generates a glyph for a single pathway by creating a bar plot of enrichment scores across specified conditions.
}
\examples{
# Example usage:
\donttest{
pathway <- "Pathway1"
conditions <- c("Condition1", "Condition2")
enrichment_scores <- data.frame(
  Condition1 = c(1.2, -0.5),
  Condition2 = c(0.8, 1.1)
)
rownames(enrichment_scores) <- c("Pathway1", "Pathway2")
enrichment_limits <- c(-1, 2)
glyph <- create_glyph_on_the_fly(pathway, conditions,
  enrichment_scores, enrichment_limits)
}
}
