% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example Enrichment Data}
\format{
A data frame with pathways as rows and conditions as columns.
}
\usage{
ex_data
}
\description{
A dataset containing example enrichment scores for pathways.
}
\details{
This dataset is used for demonstrating the package functions.
}
\keyword{datasets}
