% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph_functions.R
\name{generate_legend_plot}
\alias{generate_legend_plot}
\title{Create a enrichment score legend.}
\usage{
generate_legend_plot(conditions, enrichment_limits, reference = NULL)
}
\arguments{
\item{conditions}{Character vector of conditions to include in the glyph.}

\item{enrichment_limits}{Numeric vector of length 2 specifying the enrichment limits.}

\item{reference}{Optional string describing which conditions should be used as a reference.}
}
\value{
A ggplot object representing the legend.
}
\description{
Generates a legend of enrichment scores across specified conditions.
}
