% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_cache_directory}
\alias{get_cache_directory}
\title{Get or Create the Cache Directory}
\usage{
get_cache_directory(cache_dir = file.path(tempdir(), "glyph_cache"))
}
\arguments{
\item{cache_dir}{Character string specifying the directory to store cached glyphs. Default is \code{"glyph_cache"}.}
}
\value{
A character string representing the path to the cache directory.
}
\description{
Ensures that the specified cache directory exists. If it doesn't, the function creates it.
}
\examples{
cache_path <- get_cache_directory("my_cache")

}
