% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_child_pathways}
\alias{get_child_pathways}
\title{Get child pathways of a parent pathway}
\usage{
get_child_pathways(parent_geneset_name, mapping, g)
}
\arguments{
\item{parent_geneset_name}{Name of the parent geneset.}

\item{mapping}{Data frame containing mapping information.}

\item{g}{The pathway graph (igraph object).}
}
\value{
A character vector of child pathway names.
}
\description{
Get child pathways of a parent pathway
}
