% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_and_preprocess_gene_sets}
\alias{load_and_preprocess_gene_sets}
\title{Load and preprocess gene sets and pathway relations}
\usage{
load_and_preprocess_gene_sets(
  gene_sets,
  pathways_relation,
  translation_layer = NULL
)
}
\arguments{
\item{gene_sets}{Named list of gene sets, with pathways as names and genes as elements.}

\item{pathways_relation}{Data frame with two columns: 'parent' and 'child' pathways.}

\item{translation_layer}{Optional data frame with two columns: 'gene_sets_name' and 'relation_name'.}
}
\value{
A list containing gene sets, mapping data frame and the pathway graph.
}
\description{
Load and preprocess gene sets and pathway relations
}
