% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_roots}
\alias{reactome_roots}
\title{Reactome Roots}
\format{
A list of pathway names
}
\usage{
reactome_roots
}
\description{
A list containing the names of the top levels of the reactome pathways.
}
\details{
This list is used to connect pathways together in the plot_connected_top function
}
\keyword{datasets}
