% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph_functions.R
\name{generate_glyph_images_cached}
\alias{generate_glyph_images_cached}
\title{Generate Glyph Images with Caching}
\usage{
generate_glyph_images_cached(
  g,
  gene_sets,
  enrichment_scores,
  conditions,
  mapping,
  enrichment_limits = NULL,
  glyph_size = c(80, 60),
  res = 96,
  cache_dir = file.path(tempdir(), "glyph_cache"),
  progress = NULL,
  force_regenerate = FALSE,
  lollipop_plot = T,
  lollipop_colors = NULL
)
}
\arguments{
\item{g}{An \code{igraph} object representing the pathway graph.}

\item{gene_sets}{A named list where each element corresponds to a pathway and contains the associated genes.}

\item{enrichment_scores}{A \code{data.frame} with pathways as rows and conditions as columns, containing enrichment scores.}

\item{conditions}{A character vector specifying the conditions to visualize.}

\item{mapping}{A \code{data.frame} containing mapping information between pathways and their exact_source IDs.}

\item{enrichment_limits}{Optional numeric vector of length two specifying the minimum and maximum enrichment scores for visualization.}

\item{glyph_size}{Optional numeric vector specifying the width and height of glyphs in pixels. Default is \code{c(80, 60)}.}

\item{res}{Optional numeric value specifying the resolution of glyph images in DPI. Default is 96.}

\item{cache_dir}{Character string specifying the directory to store cached glyphs. Default is \code{"glyph_cache"}.}

\item{progress}{Optional Shiny progress object for updating progress bars.}

\item{force_regenerate}{Logical indicating whether to force regeneration of glyph images, ignoring the cache. Default is \code{FALSE}.}

\item{lollipop_plot}{Bool specifying whether we should create lollipop plots (\code{TRUE}) or not (\code{FALSE}). Lollipop plots make large numbers of comparisons easier to compare than barplots, especially when combined with lollipop_colors, with a slight loss in enrichment strength fidelity.}

\item{lollipop_colors}{Optional named list, specifies the colors of conditions in the lollipop plot. Names must match condititions exactly}
}
\value{
A named list where each name corresponds to a pathway and contains the base64-encoded image URI.
}
\description{
Generates glyph images for each pathway and caches them to improve performance.
If the enrichment scores and conditions remain unchanged, cached images are used.
}
