% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_default_data}
\alias{load_default_data}
\title{Load and preprocess gene sets}
\usage{
load_default_data()
}
\value{
A list containing gene sets, mapping data frame, exactSource to name mapping, and the pathway graph.
}
\description{
Load and preprocess gene sets
}
