% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{plot_subgraph}
\alias{plot_subgraph}
\title{Plot a Subgraph with Glyphs as Nodes}
\usage{
plot_subgraph(
  parent_geneset_name,
  enrichment_scores,
  conditions,
  mapping,
  g,
  gene_sets = NULL,
  layout = "kk",
  circular = FALSE,
  hide_nodes_without_enrichment = TRUE,
  enrichment_limits = NULL,
  use_node_label = TRUE,
  reference = NULL,
  adjust_edge_thickness = FALSE,
  edge_percentage_labels = FALSE,
  glyph_size = c(80, 60),
  res = 96,
  legend_position = NULL,
  label_wrap_width = 40,
  edge_arrows = TRUE,
  repel_labels = FALSE
)
}
\arguments{
\item{parent_geneset_name}{Name of the parent geneset. This is used to extract the subgraph of interest.}

\item{enrichment_scores}{Data frame of enrichment scores (rows are pathways, columns are conditions).}

\item{conditions}{Character vector of conditions to include in the glyphs.}

\item{mapping}{Data frame containing mapping information.}

\item{g}{The pathway graph (igraph object).}

\item{gene_sets}{List of gene sets used for edge calculations.}

\item{layout}{The layout algorithm to use (e.g., "kk", "fr"). Default is "kk".}

\item{circular}{Logical, whether to use a circular layout. Default is \code{FALSE}.}

\item{hide_nodes_without_enrichment}{Logical, whether to hide nodes without enrichment scores. Default is \code{TRUE}.}

\item{enrichment_limits}{Numeric vector of length 2 specifying the enrichment limits. If \code{NULL}, limits are calculated from the data.}

\item{use_node_label}{Logical, whether to display node labels. Default is \code{TRUE}.}

\item{reference}{Optional character string specifying the reference condition.}

\item{adjust_edge_thickness}{Logical, whether to adjust edge thickness based on overlap. Default is \code{FALSE}.}

\item{edge_percentage_labels}{Logical, whether to add edge percentage labels. Default is \code{FALSE}.}

\item{glyph_size}{Numeric vector specifying the width and height of the glyphs in pixels. Default is \code{c(80, 60)}.}

\item{res}{Numeric value specifying the resolution of the glyph images in dpi. Default is 96.}

\item{legend_position}{Optional character string ("topright", "topleft", "bottomleft" or "bottomright") determining final legend position.}

\item{label_wrap_width}{Integer specifying the maximum number of characters in each line of the label. Default is 40.}

\item{edge_arrows}{Logical, whether to display edges as arrows pointing from parents to children. Default is \code{TRUE}.}

\item{repel_labels}{Logical, whether to repel labels to avoid overlaps. Default is \code{FALSE}. If you have a very large graph it is reccommended to try this out.}
}
\value{
A ggplot object representing the subgraph.
}
\description{
Plots a subgraph with nodes represented by glyphs that visualize enrichment scores across specified conditions.
Glyphs are generated on-the-fly using the \code{create_glyph_on_the_fly} function, allowing dynamic comparisons
and efficient plotting even with a large number of conditions.
}
